backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
stripe==9.5.0
gunicorn==22.0.0
whitenoise==6.6.0
Pillow==10.3.0
requests==2.31.0
cryptography==42.0.5
celery==5.3.6
redis==5.0.4
